Changed: Refactor parser into directory module with dedicated preprocessor
- Split monolithic parser.rs into parser/mod.rs and parser/preprocessor.rs
- Moved YAML frontmatter preprocessing logic and tests to dedicated submodule
- Improved module documentation and preprocessor code readability
- Extracted first-pass detection helper for block scalar rewrite
